Split off focalplane helper utilities to their own file.  Define valid states in one place and use those to support specifying state changes with a string rather than a number.
